refactor uid check

+++ functions/index.js
@@ -265,10 +265,7 @@ exports.reportedPostsCreate = functions.region('europe-west1').firestore.documen
       const { askForHelpId, uid:uidFromDeleteRequest } = snapValue;
       const askForHelpCollectionEntry = await db.collection('/ask-for-help').doc(askForHelpId).get();
       const { uid:userIdFromCollectionEntry } = askForHelpCollectionEntry;
-      if (uidFromDeleteRequest !== userIdFromCollectionEntry) {
-        return;
-      }
-      await db.collection('/ask-for-help').doc(askForHelpId).delete()
+      if (uidFromDeleteRequest === userIdFromCollectionEntry) await db.collection('/ask-for-help').doc(askForHelpId).delete();
     } catch (e) {
       console.error(e);
       console.log('ID', snap.id);
